Add support for changing hash algorithm
Multiple allowed algorithms may be set in hawk config section, separated
by commas, the first value is the default.

Client may set algorithm in request: if algorithm is allowed in config
that will be used, else validation will fail. If no algorithm is set in
request the default from config will be used.